Final sprint 02

See merge request projeto-integrador-brasil/turma2/projeto-integrador-v1.0_equipe1!19

+++ backend/src/main/java/com/grupo01/digitalbooking/service/CategoryService.java
@@ -4,36 +4,51 @@ package com.grupo01.digitalbooking.service;
 import com.grupo01.digitalbooking.domain.Category;
 import com.grupo01.digitalbooking.dto.CategoryDTO;
 import com.grupo01.digitalbooking.repository.CategoryRepository;
-import com.grupo01.digitalbooking.service.exceptions.BadRequestException;
+import com.grupo01.digitalbooking.service.exceptions.ConflictException;
 import com.grupo01.digitalbooking.service.exceptions.NotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
+import java.util.stream.Collectors;
+
 
 @Service
 @RequiredArgsConstructor
 public class CategoryService {
 
     private final CategoryRepository categoryRepository;
+    private final String ID_CANNOT_BE_NULL = "Id não pode ser nula";
 
     public List<CategoryDTO> getCategories() {
         List<Category> categories = categoryRepository.findAll();
-        return categories.stream().map(CategoryDTO::new).toList();
+        return categories.stream().map(CategoryDTO::new).collect(Collectors.toList());
     }
 
-    public CategoryDTO getCategoryByTitle(String title) {
-        Category category = categoryRepository.findCategoryByTitle(title);
-        return new CategoryDTO(category);
+    public CategoryDTO getCategoryById(Long id) {
+
+        if (Objects.isNull(id)){
+            throw new IllegalArgumentException(ID_CANNOT_BE_NULL);
+        }
+
+        Optional<Category> category = categoryRepository.findById(id);
+
+        if (category.isPresent()) {
+            return new CategoryDTO(category.get());
+        }
+
+        throw new NotFoundException("Categoria não encontrada");
+
     }
 
     public CategoryDTO createCategory(CategoryDTO categoryDTO) {
         Category category = new Category(categoryDTO);
         Category categoryFound = (categoryRepository.findCategoryByTitle(categoryDTO.getTitle().toLowerCase()));
 
-        if(categoryFound != null){
-            throw new BadRequestException("Category " + categoryDTO.getTitle() + " already exists");
+        if(Objects.nonNull(categoryFound)){
+            throw new ConflictException("Categoria " + categoryDTO.getTitle() + " já cadastrada");
         }
 
         categoryRepository.save(category);
@@ -41,14 +56,13 @@ public class CategoryService {
     }
 
     public void deleteCategory(Long id) {
+        if (Objects.isNull(id)){
+            throw new IllegalArgumentException(ID_CANNOT_BE_NULL);
+        }
         categoryRepository.deleteById(id);
     }
 
 
-    /**
-     * TODO: Verificar se necessita regras de negocios para a edição
-     *
-     * */
     public CategoryDTO editCategory(CategoryDTO categoryDTO) {
         Category category = new Category(categoryDTO);
         Optional<Category> categoryFound = categoryRepository.findById(category.getId());
